changed CoinData last from float to BigDecimal because it's a price variable

+++ backend/src/main/java/com/wild/backend/api/btcPrice/CoinData.java
@@ -6,6 +6,8 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 
 @Getter
 @Setter
@@ -14,7 +16,7 @@ import lombok.ToString;
 public class CoinData {
 
     @JsonProperty("last")
-    private Float last;
+    private BigDecimal last;
 
     @JsonProperty("volume")
     private Float volume;
